Adicionado Teste para Data Class Statement

+++ app/src/main/java/com/github/coutinhonobre/bank/data/model/Statement.kt
@@ -4,5 +4,5 @@ data class Statement(
     val title: String,
     val desc: String,
     val date: String,
-    val value: Long
+    val value: Double
 )
